consumer read kafka topic which is relevant to outbox data

+++ product-catalog-read-service/src/main/java/com/product_catalog_read_service/consumer/DebeziumProductSyncConsumer.java
@@ -1,16 +1,25 @@
 package com.product_catalog_read_service.consumer;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.List;
+
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
 
 import com.common.avro.schemas.ProductEvent;
+import com.product_catalog_read_service.entity.ProductDocument;
 import com.product_catalog_read_service.repository.ProductRepository;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class DebeziumProductSyncConsumer {
 
     private final ProductRepository productRepository;
@@ -21,12 +30,66 @@ public class DebeziumProductSyncConsumer {
         String productId = record.key();
         ProductEvent productAvroEvent = record.value();
 
-        // 1. If Debezium sends a tombstone/null payload, remove it instantly from search visibility
+        // 1. If Debezium sends a tombstone/null payload, remove it instantly from
+        // search visibility
         if (productAvroEvent == null) {
-            productRepository.deleteById(productId);
+
+            log.info("CDC Event [HARD DELETE] -> Dropping Product ", productId);
+            if (productId != null) {
+                productRepository.deleteById(productId);
+            }
             return;
         }
 
+        if (Boolean.TRUE.equals(productAvroEvent.getIsDeleted())) {
+            System.out.println("CDC Event [SOFT DELETE] -> Dropping Product " + productAvroEvent.getId());
+            productRepository.deleteById(String.valueOf(productAvroEvent.getId()));
+            return;
+        }
+
+        log.info("CDC Event [SYNC] -> Processing Index Update for Product: ", productAvroEvent.getId());
+
+        saveDocument(productAvroEvent);
+
+    }
+
+    private void saveDocument(ProductEvent productAvroEvent) {
+
+        // 3. Map Avro fields to your Elasticsearch Document
+        ProductDocument doc = new ProductDocument();
+        doc.setId(String.valueOf(productAvroEvent.getId()));
+        doc.setEventId(productAvroEvent.getEventId().toString());
+        doc.setSku(productAvroEvent.getSku().toString());
+        doc.setName(productAvroEvent.getName().toString());
+
+        // Handle nullable description
+        doc.setDescription(
+                productAvroEvent.getDescription() != null ? productAvroEvent.getDescription().toString() : "");
+
+        doc.setBrandId(productAvroEvent.getBrandId());
+        doc.setBrandName(productAvroEvent.getBrandName().toString());
+        doc.setCategoryId(productAvroEvent.getCategoryId());
+        doc.setCategoryName(productAvroEvent.getCategoryName().toString());
+
+        // Safe conversion of Avro Decimal logic layer to Java BigDecimal
+        // BigDecimal extractedPrice = productAvroEvent.getPrice();
+
+        java.nio.ByteBuffer buffer = productAvroEvent.getPrice();
+        byte[] bytes = new byte[buffer.remaining()];
+        buffer.get(bytes);
+
+        BigDecimal extractedPrice = new BigDecimal(new BigInteger(bytes), 2);
+        doc.setPrice(extractedPrice);
+
+        doc.setStockCount(productAvroEvent.getStockCount());
+        doc.setStatus(productAvroEvent.getStatus().toString());
+        doc.setIsDeleted(productAvroEvent.getIsDeleted());
+
+        Instant timestampInstant = productAvroEvent.getUpdatedAt();
+        doc.setUpdatedAt(timestampInstant.toEpochMilli());
+
+        // 4. Commit directly into Elasticsearch index
+        productRepository.save(doc);
     }
 
 }
